requirements.txt
customtkinter==5.2.2
cvzone==1.6.1
easyocr==1.7.2
numpy==1.25.2
opencv_contrib_python==4.11.0.86
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
Pillow==11.2.1
ultralytics==8.3.111


